chore(apps/interface): refactor themes
