+++ apps/interface/tests/utils/renderApp.tsx
@@ -1,7 +1,7 @@
 import React, { FC, ReactElement } from "react";
 import { render, RenderOptions } from "@testing-library/react";
 import { ChakraProvider } from "@chakra-ui/react";
-import theme from "../../utils/theme";
+import themes from "../../themes";
 
 // Rainbowkit
 import { RainbowKitProvider } from "@rainbow-me/rainbowkit";
@@ -23,7 +23,7 @@ const wagmiClient = createClient({
 
 const Wrapper: FC<{ children: React.ReactNode }> = ({ children }) => {
     return (
-        <ChakraProvider theme={theme}>
+        <ChakraProvider theme={themes}>
             <WagmiConfig client={wagmiClient}>{children}</WagmiConfig>
         </ChakraProvider>
     );
